FinBIF habitat types

Display the habitat types and specific habitat types used in the FinBIF database.

finbif_habitat_types()

finbif_habitat_qualifiers()

Value

A data.frame.

Examples

# Display the administrative statuses used by FinBIF
finbif_habitat_types()

#>                                                       habitat_type code
#> 1                                                          Forests    M
#> 2                                                    Heath forests   MK
#> 3                        Sub-xeric, xeric and barren heath forests  MKK
#> 4                                Mesic and herb-rich heath forests  MKT
#> 5                        Herb-rich forests (also spruce-dominated)   ML
#> 6                                  Dry and mesic herb-rich forests  MLT
#> 7                                          Moist herb-rich forests  MLK
#> 8  Alpine birch forests (excluding herb-rich alpine birch forests)   MT
#> 9                                   Herb-rich alpine birch forests  MTL
#> 10                                                           Mires    S
#> 11                                                       Rich fens   SL
#> 12                                                  Open rich fens  SLA
#> 13                                                  Rich pine fens  SLR
#> 14                                          Rich spruce-birch fens  SLK
#> 15                                                            Fens   SN
#> 16                                    Ombro- and oligotrophic fens  SNK
#> 17                                                Mesotrophic fens  SNR
#> 18                                                      Pine mires   SR
#> 19                              Ombro- and oligotrophic pine mires  SRK
#> 20                                          Mesotrophic pine mires  SRR
#> 21                                                    Spruce mires   SK
#> 22                                       Oligotrophic spruce mires  SKK
#> 23                          Eutrophic and mesotrophic spruce mires  SKR
#> 24                                                Aquatic habitats    V
#> 25                                                      Baltic sea   VI
#> 26            Benthic habitats characterized by rocks and boulders  VIK
#> 27                  Benthic habitats characterized by mud and silt  VIM
#> 28                        Benthic habitats characterized by gravel  VIS
#> 29                          Benthic habitats characterized by sand  VIH
#> 30                                          Mixed benthic habitats  VIE
#> 31                                                Pelagic habitats  VIP
#> 32                                                 Lakes and ponds   VS
#> 33                                    Oligotrophic lakes and ponds  VSK
#> 34                       Eutrophic and mesotrophic lakes and ponds  VSR
#> 35                                Small ponds (also in mires etc.)   VA
#> 36                                                          Rivers   VJ
#> 37                                              Brooks and streams   VP
#> 38                                                          Rapids   VK
#> 39                                                Spring complexes   VL
#> 40                                                          Shores    R
#> 41                                                  Coastal shores   RI
#> 42                                                 Coastal forests  RIM
#> 43                                         Coastal flooded forests RIMT
#> 44                                           Coastal wooded swamps RIML
#> 45                                            Coastal shore scrubs  RIP
#> 46                                                 Coastal meadows  RIN
#> 47                                            Coastal swamp shores  RIL
#> 48                      Coastal reedbeds with phragmites australis  RIR
#> 49                    Coastal gravel-, shingle- and boulder shores  RIS
#> 50                                             Coastal sand shores  RIH
#> 51                                    Coastal open alluvial shores  RIT
#> 52                                     Lake shores and river banks   RJ
#> 53                   Inland forests on lake shores and river banks  RJM
#> 54                                          Inland flooded forests RJMT
#> 55                                            Inland wooded swamps RJML
#> 56                                             Inland shore scrubs  RJP
#> 57                          Meadows on lake shores and river banks  RJN
#> 58                                             Inland swamp shores  RJL
#> 59                       Inland reedbeds with phragmites australis  RJR
#> 60                     Inland gravel-, shingle- and boulder shores  RJS
#> 61                               Sandy lake shores and river banks  RJH
#> 62                                     Inland open alluvial shores  RJT
#> 63                                Rock outcrops and boulder fields    K
#> 64                           Calcareous rock outcrops and quarries   KK
#> 65                                        Serpentine rock outcrops   KS
#> 66                                              Canyons and gorges   KR
#> 67                                              Caves and crevices   KL
#> 68                                         Siliceous rock outcrops   KM
#> 69                                                 Alpine habitats    T
#> 70                                                   Alpine heaths   TK
#> 71                                                  Alpine meadows   TN
#> 72                                                        Snowbeds   TU
#> 73                                             Alpine heath scrubs   TP
#> 74                         Alpine rock outcrops and boulder fields   TL
#> 75               Alpine siliceous rock outcrops and boulder fields  TLL
#> 76              Alpine calcareous rock outcrops and boulder fields  TLK
#> 77              Alpine serpentine rock outcrops and boulder fields  TLS
#> 78                                       Alpine gorges and canyons  TLR
#> 79                       Alpine spring complexes and seepage areas  TLÄ
#> 80                                                    Alpine mires   TS
#> 81                        Alpine ponds and lakes (shores included)   TJ
#> 82                               Paljakan virtavedet (sis. rannat)   TV
#> 83                                   Paljakan lampareet ja allikot   TA
#> 84                            Rural biotopes and cultural habitats    I
#> 85                                         Seminatural dry meadows   IN
#> 86                                       Seminatural mesic meadows   IT
#> 87           Wooded pastures, pollard meadows and grazed woodlands   IH
#> 88             Seminatural moist meadows (excluding shore meadows)   IK
#> 89                                                    Ditches etc.   IO
#> 90                                                     Arable land   IV
#> 91                                  Puistot, pihamaat ja puutarhat   IP
#> 92    Open habitats formed and maintained by heavy human influence   IU
#> 93                                     Buildings and constructions   IR
#> 94                                                 Habitat unknown    ?

finbif_habitat_qualifiers()

#>                                                                                 habitat_type
#> 1  Old-growth forests (including sporadic old trees and younger forests with decaying trees)
#> 2                                                      Esker forests, also semi-open forests
#> 3                            Burnt forest areas and other young stages of natural succession
#> 4                                                        Broadleaved deciduous trees present
#> 5                                                                                Sun-exposed
#> 6                                                                                      Shady
#> 7            On the shore (aquatic habitat is essential for the distribution of the species)
#> 8                                                            Acidic rocks and boulder fields
#> 9                                        Intermediate-basic rock outcrops and boulder fields
#> 10                                                                         Calcareous effect
#>    code
#> 1     V
#> 2     H
#> 3     P
#> 4     J
#> 5    PA
#> 6    VA
#> 7    RA
#> 8    KA
#> 9    KE
#> 10   CA