Get FinBIF records

Download records from FinBIF.

finbif_records(
  filter,
  select,
  order_by,
  sample = FALSE,
  n = 10,
  page = 1,
  count_only = FALSE,
  quiet = FALSE,
  cache = getOption("finbif_use_cache"),
  seed
)

Arguments

filter
List of named character vectors. Filters to apply to records.

select
Character vector. Variables to return. If not specified a default set of commonly used variables will be used. Use "default_vars" as a shortcut for this set.

order_by
Character vector. Variables to order records by before they are returned. Most, though not all, variables can be used to order records before they are returned. Ordering is ascending by default. To return in descending order append a - to the front of the variable (e.g., "-date_start"). Default order is "-date_start" > "-load_data" > "reported_name".

sample
Logical. If TRUE randomly sample the records from the FinBIF database.

n
Integer. How many records to download.

page
Integer. Which page of records to start downloading from.

count_only
Logical. Only return the number of records available.

quiet
Logical. Suppress the progress indicator for multipage downloads.

cache
Logical. Use cached data.

seed
Integer. Set a seed for randomly sampling records. Note that the server currently ignores seed setting and this argument has currently has little effect.

Value

A finbif_api or finbif_api_list object.

Examples

if (FALSE) {

# Get the last 100 records from FinBIF
finbif_records(n = 100)
}